Precise VsyncRateMin calculation
This change adds a precise calculation of the minimum supported vsync
rate, replacing the value previously accessed from the EDID which is
imprecise due to being encoded as an integer. The precise value is
determined in DisplayMode, and the type is updated accordingly from int
to float where relevant.

In order to support the precise vsync_min calculation, detailed timing
fields are added to DisplayMode which are taken directly from the kernel
mode object. This change is aligned with the goal of improving mode
matching across kernel and userspace (which was not previously possible
when modes had equal resolutions and rates). In future changes,
additional kernel mode fields can be added to DisplayMode as needed.

Bug: b:297556474
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4811479
Commit-Queue: Andrew Wolfers <aswolfers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1202140}

+++ ui/display/manager/display_manager.cc
@@ -640,7 +640,7 @@ void DisplayManager::RegisterDisplayProperty(
     float refresh_rate,
     bool is_interlaced,
     VariableRefreshRateState variable_refresh_rate_state,
-    const absl::optional<uint16_t>& vsync_rate_min) {
+    const absl::optional<float>& vsync_rate_min) {
   if (display_info_.find(display_id) == display_info_.end())
     display_info_[display_id] =
         ManagedDisplayInfo(display_id, std::string(), false);
