+++ ui/display/manager/managed_display_info.h
@@ -333,10 +333,10 @@ class DISPLAY_MANAGER_EXPORT ManagedDisplayInfo {
     variable_refresh_rate_state_ = variable_refresh_rate_state;
   }
 
-  const absl::optional<uint16_t>& vsync_rate_min() const {
+  const absl::optional<float>& vsync_rate_min() const {
     return vsync_rate_min_;
   }
-  void set_vsync_rate_min(const absl::optional<uint16_t>& vsync_rate_min) {
+  void set_vsync_rate_min(const absl::optional<float>& vsync_rate_min) {
     vsync_rate_min_ = vsync_rate_min;
   }
 
@@ -462,7 +462,7 @@ class DISPLAY_MANAGER_EXPORT ManagedDisplayInfo {
   DrmFormatsAndModifiers drm_formats_and_modifiers_;
 
   VariableRefreshRateState variable_refresh_rate_state_;
-  absl::optional<uint16_t> vsync_rate_min_;
+  absl::optional<float> vsync_rate_min_;
 
   // If you add a new member, you need to update Copy().
 };
